requirements.txt
flask
flask-mysql-connector
flask-jwt-extended
bcrypt
python-dotenv